Module Prometheus_unix

Report metrics for Prometheus. See: https://prometheus.io/

Notes:

This module is intended to be used by applications that export Prometheus metrics. Libraries should only link against the `Prometheus` module.

This module automatically initialises itself and registers some standard collectors relating to GC statistics, as recommended by Prometheus.

This extends Prometheus_app with support for cmdliner option parsing, a server pre-configured for Unix, and a start-time metric that uses Unix.gettimeofday.

type config

val serve : config -> unit Lwt.t list

serve config starts a Cohttp server according to config. It returns a singleton list containing the thread to monitor, or an empty list if no server is configured.

val opts : config Cmdliner.Term.t

opts is the extra command-line options to offer Prometheus monitoring.

module Logging : sig ... end

Report metrics for messages logged.